MockAltNetwork: move to o.b.base.internal
